consolidate camerapositionerwindow with gui classes
